% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDC.R
\name{ALDC}
\alias{ALDC}
\title{Average Levenshtein Distance of Candidates - ALDC}
\usage{
ALDC(prod_acc, data)
}
\arguments{
\item{prod_acc}{A list, typically the output of \code{WpmWithLdl::accuracy_production}.}

\item{data}{The dataset with which \code{prod_acc} was computed}
}
\value{
A data frame containing:
\itemize{
\item \code{Word} - The pertinent word as found in \code{data}.
\item \code{Base} - The pertinent base as found in \code{data}.
\item \code{ALDC} - The ALDC of a word.
}
}
\description{
\code{ALDC} computes the mean Levenshtein Distances of all candidate productions.
}
\references{
Baayen, R. H., Chuang, Y. Y., and Blevins, J. P. (2018). Inflectional morphology with linear mappings. The Mental Lexicon, 13 (2), 232-270.

Chuang, Y-Y., Vollmer, M-l., Shafaei-Bajestan, E., Gahl, S., Hendrix, P., & Baayen, R. H. (2020). The processing of pseudoword form and meaning in production and comprehension: A computational modeling approach using Linear Discriminative Learning. Behavior Research Methods, 1-51.

Levenshtein, V. I. (1966). Binary codes capable of correcting deletions, insertions, and reversals. Soviet Physics Doklady. 10 (8), 707-710.

van der Loo M (2014). The stringdist package for approximate string matching. The R Journal, 6, 111-122.
}
\author{
D. Schmitz
}
