% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRC.R
\name{DRC}
\alias{DRC}
\title{Dual Route Consistency - DRC}
\usage{
DRC(comprehension, Shat_matrix_indirect, data)
}
\arguments{
\item{comprehension}{A list, typically computed by solving \code{WpmWithLdl::learn_comprehension}.}

\item{Shat_matrix_indirect}{A matrix, typically computed by solving \code{WpmWithLdl::learn_comprehension} with \code{indirectRoute = TRUE}.}

\item{data}{The dataset with which \code{comprehension} and \code{Shat_matrix_indirect} were computed.}
}
\value{
A data frame containing:
\itemize{
\item \code{DRC} - The DRC value of a word.
\item \code{Word} - The pertinent word as found in \code{data}.
\item \code{Base} - The pertinent base as found in \code{data}.
}
}
\description{
\code{DRC} computes the correlation between the semantic vector estimated from the direct route
and that from the indirect route.
}
\references{
Baayen, R. H., Chuang, Y. Y., and Blevins, J. P. (2018). Inflectional morphology with linear mappings. The Mental Lexicon, 13 (2), 232-270.

Chuang, Y-Y., Vollmer, M-l., Shafaei-Bajestan, E., Gahl, S., Hendrix, P., & Baayen, R. H. (2020). The processing of pseudoword form and meaning in production and comprehension: A computational modeling approach using Linear Discriminative Learning. Behavior Research Methods, 1-51.
}
\author{
D. Schmitz
}
