% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDNN.R
\name{EDNN}
\alias{EDNN}
\title{Euclidian Distance from Nearest Neighbour - EDNN}
\usage{
EDNN(comprehension, data)
}
\arguments{
\item{comprehension}{A data frame, typically the output of \code{WpmWithLdl::learn_comprehension}.}

\item{data}{The dataset with which \code{comprehension} was computed}
}
\value{
A data frame containing:
\itemize{
\item \code{EDNN} - The EDNN value of a word.
\item \code{Word} - The pertinent word as found in \code{data}.
\item \code{Base} - The pertinent base as found in \code{data}.
}
}
\description{
\code{EDNN} extracts the Euclidian distance between a word's estimated semantic vector and its nearest neighbour.

\code{EDNN} extracts the Euclidian distance between a word's estimated semantic vector and its nearest neighbour.
}
\references{
Baayen, R. H., Chuang, Y. Y., and Blevins, J. P. (2018). Inflectional morphology with linear mappings. The Mental Lexicon, 13 (2), 232-270.

Beygelzimer, A., Kakadet, S., Langford, J., Arya, S., Mount, D., & Shengqiao, L. (2019). FNN: Fast Nearest Neighbor Search Algorithms and Applications. R package version 1.1.3.

Chuang, Y-Y., Vollmer, M-l., Shafaei-Bajestan, E., Gahl, S., Hendrix, P., & Baayen, R. H. (2020). The processing of pseudoword form and meaning in production and comprehension: A computational modeling approach using Linear Discriminative Learning. Behavior Research Methods, 1-51.
}
\author{
D. Schmitz
}
