% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDLConvFunctions_data.R
\docType{data}
\name{LDLConvFunctions_data}
\alias{LDLConvFunctions_data}
\title{LDLConvFunction - Data Set}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{Word}{Words - real and nonce words}
\item{DISC}{Transcription of Word}
\item{Affix}{Affix of the word, if applicable}
\item{Base}{Base of the word}
}
}
\usage{
LDLConvFunctions_data
}
\description{
Small data set containing English real words as well as pseudowords (from Schmitz et al., 2020). Columns contain information on wordforms, bases, affixes, and transcription.
}
\references{
Schmitz, D., Baer-Henney, D., & Plag, I. (submitted). The duration of word-final /s/ differs across morphological categories in English: Evidence from pseudowords.
}
\author{
D. Schmitz
}
\keyword{datasets}
