% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWS.R
\name{MWS}
\alias{MWS}
\title{Mean Word Support - MWS}
\usage{
MWS(prod_acc, prod_measures)
}
\arguments{
\item{prod_acc}{A list, typically the output of \code{WpmWithLdl::accuracy_production}.}

\item{prod_measures}{A data frame, typically the output of \code{WpmWithLdl::production_measures}.}
}
\value{
A data frame containing:
\itemize{
\item \code{MWS} - The mean word support.
}
}
\description{
\code{MWS} computes the mean word support of a word's predicted form. Note: MWS only takes into account the winning production form.
It basically takes \code{path_sum} and divides it by the number of paths.
}
\references{
Baayen, R. H., Chuang, Y. Y., and Blevins, J. P. (2018). Inflectional morphology with linear mappings. The Mental Lexicon, 13 (2), 232-270.

Stein, S., & Plag, I. (2021). Modeling affix durations with Linear Discriminative Learning. \link{Unpublished Manuscript}.
}
\author{
D. Schmitz
}
