% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_pseudo_C_matrix.R
\name{reorder_pseudo_C_matrix}
\alias{reorder_pseudo_C_matrix}
\title{reorder_pseudo_C_matrix}
\usage{
reorder_pseudo_C_matrix(pseudo_C_matrix, real_C_matrix, found_triphones)
}
\arguments{
\item{pseudo_C_matrix}{A data frame, typically the output of \code{WpmWithLdl::make_cue_matrix}.}

\item{real_C_matrix}{A data frame, typically the output of \code{WpmWithLdl::make_cue_matrix}.}

\item{found_triphones}{A data frame, typically the output of \code{find_triphones}.}
}
\value{
A cue matrix (C) similar to those found in the result of \code{WpmWithLdl::make_cue_matrix}.
}
\description{
\code{reorder_pseudo_C_matrix} reorders and extends a smaller C matrix according to the order and number of columns of a bigger C matrix.
Use \code{find_triphones} to provide information on column order.
}
\references{
Baayen, R. H., Chuang, Y. Y., and Blevins, J. P. (2018). Inflectional morphology with linear mappings. The Mental Lexicon, 13 (2), 232-270.
}
\author{
D. Schmitz
}
