% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDNN.R
\name{EDNN}
\alias{EDNN}
\title{Euclidian Distance from Nearest Neighbour - EDNN}
\usage{
EDNN(comp_measures, data)
}
\arguments{
\item{comp_measures}{A data frame, typically the output of \code{WpmWithLdl::comprehension_measures}.}

\item{data}{The dataset with which \code{comp_measures} was computed}
}
\value{
A data frame containing:
\itemize{
\item \code{EDNN} - The EDNN value of a word.
\item \code{Word} - The pertinent word as found in \code{data}.
\item \code{Base} - The pertinent base as found in \code{data}.
}
}
\description{
\code{EDNN} extracts the Euclidian distance between a word's estimated semantic vector and its nearest neighbour.
This measure is similar to \code{l2norm} which is part of the \code{WpmWithLdl::comprehension_measures} function.
}
\references{
Baayen, R. H., Chuang, Y. Y., and Blevins, J. P. (2018). Inflectional morphology with linear mappings. The Mental Lexicon, 13 (2), 232-270.

Chuang, Y-Y., Vollmer, M-l., Shafaei-Bajestan, E., Gahl, S., Hendrix, P., & Baayen, R. H. (2020). The processing of pseudoword form and meaning in production and comprehension: A computational modeling approach using Linear Discriminative Learning. Behavior Research Methods, 1-51.
}
\author{
D. Schmitz
}
